Add ClearAllTiles button and update GroundTileData instantiation
Added a new method ClearAllTiles() in TilemapManager.cs to allow clearing all tiles from the baseTilemap. Also, updated the instantiation of GroundTileData to set the WhatIsOnIt enum based on the type of tile (Building or Water) instead of the boolean isTilePopulated. This ensures proper categorization of tiles in the GroundTileData.